+++ packages/vue-2-password/src/components/Password-ts.vue
@@ -131,7 +131,7 @@ export default Vue.extend({
         return {
             id: nanoid(6),
             type: "password",
-            empty: this.value.length == 0,
+            empty: true,
         };
     },